feat: QueryBoundary 추가 및 디렉토리 위치 변경
