+++ frontend/src/components/@common/ErrorBoundary/QueryBoundary/QueryBoundary.tsx
@@ -1,0 +1,39 @@
+import { QueryErrorResetBoundary, useQueryErrorResetBoundary } from '@tanstack/react-query';
+import { ComponentProps, PropsWithChildren, ReactNode, Suspense } from 'react';
+
+import { composeFunctions } from '@/utils/composeFunctions';
+import { composeEventHandlers } from '@/utils/dom';
+
+import LoadingSpinner from '../../LoadingSpinner';
+import APIBoundary from '../APIBoundary/APIBoundary';
+
+interface QueryBoundaryProps extends Omit<ComponentProps<typeof APIBoundary>, 'fallback'> {
+  errorFallback?: ComponentProps<typeof APIBoundary>['fallback'];
+  loadingFallback?: ComponentProps<typeof Suspense>['fallback'];
+}
+
+const QueryBoundary = (props: PropsWithChildren<QueryBoundaryProps>) => {
+  const {
+    children,
+    loadingFallback = <LoadingSpinner />,
+    errorFallback,
+    onReset,
+    ...restProps
+  } = props;
+
+  return (
+    <QueryErrorResetBoundary>
+      {({ reset }) => (
+        <APIBoundary
+          fallback={errorFallback}
+          onReset={composeFunctions(reset, onReset)}
+          {...restProps}
+        >
+          <Suspense fallback={loadingFallback}>{children}</Suspense>
+        </APIBoundary>
+      )}
+    </QueryErrorResetBoundary>
+  );
+};
+
+export default QueryBoundary;
